+++ flake.nix
@@ -2,12 +2,14 @@
   description = "Shrun is a tool for concurrently running shell commands.";
   inputs = {
     algebra-simple-src.url = "github:tbidne/algebra-simple";
+    byte-types-src.url = "github:tbidne/byte-types";
     flake-utils.url = "github:numtide/flake-utils";
     nixpkgs.url = "github:nixos/nixpkgs/nixos-unstable";
     relative-time-src.url = "github:tbidne/relative-time";
   };
   outputs =
     { algebra-simple-src
+    , byte-types-src
     , flake-utils
     , nixpkgs
     , relative-time-src
@@ -33,6 +35,8 @@
           overrides = final: prev: with compiler; {
             algebra-simple =
               final.callCabal2nix "algebra-simple" algebra-simple-src { };
+            byte-types =
+              final.callCabal2nix "byte-types" byte-types-src { };
             package-version = pkgs.haskell.lib.doJailbreak prev.package-version;
             relative-time =
               final.callCabal2nix "relative-time" relative-time-src { };